Move ManifestWork into its own utils
Refactoring drpc controller to split into utils.

Moving MW related functions into it's own utils.
