+++ controllers/util/mw_util_test.go
@@ -1,0 +1,217 @@
+package util_test
+
+import (
+	"time"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+	ocmworkv1 "github.com/open-cluster-management/api/work/v1"
+	rmnutil "github.com/ramendr/ramen/controllers/util"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+var _ = Describe("DRPlacementControl Reconciler", func() {
+	Context("IsManifestInAppliedState checks ManifestWork with single timestamp", func() {
+		timeOld := time.Now().Local()
+		timeMostRecent := timeOld.Add(time.Second)
+
+		It("'Applied' present, Status False", func() {
+			mw := &ocmworkv1.ManifestWork{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "ramendr-vrg-roles",
+				},
+				Status: ocmworkv1.ManifestWorkStatus{
+					Conditions: []metav1.Condition{
+						{
+							Type:               ocmworkv1.WorkApplied,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+							Status:             metav1.ConditionFalse,
+						},
+					},
+				},
+			}
+
+			Expect(rmnutil.IsManifestInAppliedState(mw)).To(Equal(false))
+		})
+
+		It("'Applied' present, Status true", func() {
+			mw := &ocmworkv1.ManifestWork{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "ramendr-vrg-roles",
+				},
+				Status: ocmworkv1.ManifestWorkStatus{
+					Conditions: []metav1.Condition{
+						{
+							Type:               ocmworkv1.WorkApplied,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+							Status:             metav1.ConditionTrue,
+						},
+					},
+				},
+			}
+
+			Expect(rmnutil.IsManifestInAppliedState(mw)).To(Equal(true))
+		})
+
+		It("'Applied' not present", func() {
+			mw := &ocmworkv1.ManifestWork{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "ramendr-vrg-roles",
+				},
+				Status: ocmworkv1.ManifestWorkStatus{
+					Conditions: []metav1.Condition{
+						{
+							Type:               ocmworkv1.WorkProgressing,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+						},
+					},
+				},
+			}
+
+			Expect(rmnutil.IsManifestInAppliedState(mw)).To(Equal(false))
+		})
+	})
+
+	Context("IsManifestInAppliedState checks ManifestWork with multiple timestamps", func() {
+		timeOld := time.Now().Local()
+		timeMostRecent := timeOld.Add(time.Second)
+
+		It("no duplicate timestamps, 'Applied' is most recent", func() {
+			mw := &ocmworkv1.ManifestWork{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "ramendr-vrg-roles",
+				},
+				Status: ocmworkv1.ManifestWorkStatus{
+					Conditions: []metav1.Condition{
+						{
+							Type:               ocmworkv1.WorkApplied,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+							Status:             metav1.ConditionTrue,
+						},
+						{
+							Type:               ocmworkv1.WorkProgressing,
+							LastTransitionTime: metav1.Time{Time: timeOld},
+						},
+					},
+				},
+			}
+
+			Expect(rmnutil.IsManifestInAppliedState(mw)).To(Equal(true))
+		})
+
+		It("no duplicates timestamps, 'Applied' is not most recent", func() {
+			mw := &ocmworkv1.ManifestWork{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "ramendr-vrg-roles",
+				},
+				Status: ocmworkv1.ManifestWorkStatus{
+					Conditions: []metav1.Condition{
+						{
+							Type:               ocmworkv1.WorkProgressing,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+						},
+						{
+							Type:               ocmworkv1.WorkApplied,
+							LastTransitionTime: metav1.Time{Time: timeOld},
+						},
+					},
+				},
+			}
+
+			Expect(rmnutil.IsManifestInAppliedState(mw)).To(Equal(false))
+		})
+
+		It("with duplicate timestamps, 'Applied' is most recent", func() {
+			mw := &ocmworkv1.ManifestWork{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "ramendr-vrg-roles",
+				},
+				Status: ocmworkv1.ManifestWorkStatus{
+					Conditions: []metav1.Condition{
+						{
+							Type:               ocmworkv1.WorkApplied,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+							Status:             metav1.ConditionTrue,
+						},
+						{
+							Type:               ocmworkv1.WorkProgressing,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+							Status:             metav1.ConditionUnknown,
+						},
+						{
+							Type:               ocmworkv1.WorkProgressing,
+							LastTransitionTime: metav1.Time{Time: timeOld},
+						},
+					},
+				},
+			}
+
+			Expect(rmnutil.IsManifestInAppliedState(mw)).To(Equal(true))
+		})
+
+		It("with duplicate timestamps, 'Applied' is not most recent", func() {
+			mw := &ocmworkv1.ManifestWork{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "ramendr-vrg-roles",
+				},
+				Status: ocmworkv1.ManifestWorkStatus{
+					Conditions: []metav1.Condition{
+						{
+							Type:               ocmworkv1.WorkAvailable,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+						},
+						{
+							Type:               ocmworkv1.WorkProgressing,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+						},
+						{
+							Type:               ocmworkv1.WorkApplied,
+							LastTransitionTime: metav1.Time{Time: timeOld},
+						},
+					},
+				},
+			}
+
+			Expect(rmnutil.IsManifestInAppliedState(mw)).To(Equal(false))
+		})
+
+		It("duplicate timestamps with Degraded and Applied status", func() {
+			mw := &ocmworkv1.ManifestWork{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "ramendr-vrg-roles",
+				},
+				Status: ocmworkv1.ManifestWorkStatus{
+					Conditions: []metav1.Condition{
+						{
+							Type:               ocmworkv1.WorkApplied,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+						},
+						{
+							Type:               ocmworkv1.WorkDegraded,
+							LastTransitionTime: metav1.Time{Time: timeMostRecent},
+						},
+					},
+				},
+			}
+
+			Expect(rmnutil.IsManifestInAppliedState(mw)).To(Equal(false))
+		})
+	})
+
+	Context("IsManifestInAppliedState checks ManifestWork with no timestamps", func() {
+		It("manifest missing conditions", func() {
+			mw := &ocmworkv1.ManifestWork{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "ramendr-vrg-roles",
+				},
+				Status: ocmworkv1.ManifestWorkStatus{
+					Conditions: []metav1.Condition{
+						// empty
+					},
+				},
+			}
+
+			Expect(rmnutil.IsManifestInAppliedState(mw)).To(Equal(false))
+		})
+	})
+})
